Implemented most of Vida's comments.
I didn't do these things, or only partially did them:

The following things I didn't do:

> * again ranges of r and k are missing from most lemmans/thms in the paper

I added this in most places, but a few of the lemmas/theorems are not self-contained (they refer to things in earlier paragraphs). In these cases I added any ranges in those paragraphs.

> * Btw. we should mention the for k=1 case, the upper and lower bound are
> trivially r-1 and that thus we consider k>1 case.

I wasn't sure where to put this and, right now, I can't add a single line to the paper without overflowing it.  Again, I'll see if I can fit it in later.

> * Line 18. Add references to support this claim or say that we will show
> this later if that is the case.

The full paper has a figure that shows a simple example.  I don't know of a reference and there's no room for the figure.  I'll see what I can do once we get down to the wire.

> * In the references, in some places we have only abriviations for
> conferences like SODA and CCCG, in other we have a full name of conferences
> as well (preferred option, for me at least), like GD.

I'm saving this as the last thing I do, if there's still time.